+++ changelog.md
@@ -8,7 +8,8 @@
 ##### Misc
 - Change how to specify custom keybindings
 - 'yank_to_clipboard' now belongs to mode 'global' (before: 'normal')
-- Rename 'normal' mode to 'navigate' mode
+- Rename 'normal mode' to 'navigate mode'
+- Rename 'edit mode' to 'operate mode'
 
 ## v0.16.1: July 03, 2018
 ##### Features